pyproject.toml
[project]
name = "vinas-external-api-app"
version = "0.1.0"
description = ""
authors = [
    {name = "peyro",email = "peyrovskaaa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "redis (>=6.2.0,<7.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "fastapi-problem (>=0.11.4,<0.12.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "alembic (>=1.16.4,<2.0.0)",
]

[tool.poetry]
packages = [{include = "vinas_external_api_app", from = "src"}]


[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.21.0"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
